requirements.txt
astropy==3.2.3
dominate==2.6.0
imageio==2.9.0
ipython==8.11.0
kornia==0.5.0
matplotlib==3.1.2
numpy==1.16.4
opencv_python==3.4.2.16
pandas==1.1.3
Pillow==9.4.0
pycocotools==2.0.4
PyYAML==6.0
rawpy==0.18.0
scikit_image==0.16.2
scipy==1.2.1
skimage==0.0
tensorboardX==2.6
thop==0.1.1.post2209072238
timm==0.3.2
torch==1.8.1+cu101
torchprofile==0.0.4
torchsummary==1.5.1
torchvision==0.9.1+cu101
tqdm==4.51.0
visdom==0.2.4
yacs==0.1.8
